Smart_Data_Mapper/pyproject.toml
[project]
name = "smart-data-mapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "numpy>=2.3.2",
    "openai>=1.97.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.1",
    "streamlit>=1.47.1",
]
